% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xl_etiquetas.R
\name{xl_etiquetas}
\alias{xl_etiquetas}
\title{Tabela de etiquetas de excel}
\usage{
xl_etiquetas(arquivo)
}
\arguments{
\item{arquivo}{O caminho para uma planilha de excel contendo etiquetas}
}
\value{
Uma tabela (tibble)
}
\description{
Cria uma tabela (tibble) a partir das etiquetas contidas em um arquivo de excel.
Para planilhas de labels da Innovare no formato "Var/Val", acessa a aba "Val" e coloca as etiquetas na primeira coluna, os códgios na segunda coluna e os labels na terceira coluna
}
\examples{
planilha <- system.file("extdata", "planilha.xlsx", package = "innoutils")
xl_etiquetas(arquivo = planilha)
}
